v7.0.5

See merge request scotch/scotch!22

+++ src/libscotch/bdgraph_bipart_df.c
@@ -1,4 +1,4 @@
-/* Copyright 2007,2008,2011,2014,2023 IPB, Universite de Bordeaux, INRIA & CNRS
+/* Copyright 2007,2008,2011,2014,2023,2024 IPB, Universite de Bordeaux, INRIA & CNRS
 **
 ** This file is part of the Scotch software package for static mapping,
 ** graph partitioning and sparse matrix ordering.
@@ -46,7 +46,7 @@
 /**                # Version 6.0  : from : 11 sep 2011     **/
 /**                                 to   : 31 aug 2014     **/
 /**                # Version 7.0  : from : 17 jan 2023     **/
-/**                                 to   : 17 jan 2023     **/
+/**                                 to   : 09 aug 2024     **/
 /**                                                        **/
 /************************************************************/
 
@@ -213,8 +213,8 @@ const BdgraphBipartDfParam * const  paraptr)      /*+ Method parameters +*/
   }
 
   ovflval    = 0;
-  cdifval    = paraptr->cdifval;
-  cremval    = paraptr->cremval;
+  cdifval    = (float) paraptr->cdifval;
+  cremval    = (float) paraptr->cremval;
   edgegsttax = grafptr->s.edgegsttax;
   for (passnum = 0; ; ) {                         /* For all passes         */
     if (ovflval == 0) {                           /* If no overflow occured */
